s3dm-mvp/requirements.txt
fastapi
uvicorn[standard]
pymongo
httpx
transformers
torch # Or tensorflow if you prefer TF models for LLM
pydantic
python-dotenv
pycryptodome